Misc Improvements + respecting post visibility state when listing threadmarks

+++ upload/library/Sidane/Threadmarks/Install.php
@@ -4,35 +4,59 @@ class Sidane_Threadmarks_Install
 {
   public static function install($existingAddOn, $addOnData)
   {
-    if ($existingAddOn['version_id'] > 1) {
-      return;
-    }
+    $version = isset($existingAddOn['version_id']) ? $existingAddOn['version_id'] : 0;
 
     $db = XenForo_Application::get('db');
+    $tables_created = false;
 
-    $db->query("
-      CREATE TABLE IF NOT EXISTS threadmarks (
-        threadmark_id INT UNSIGNED NOT NULL PRIMARY KEY AUTO_INCREMENT,
-        thread_id INT UNSIGNED NOT NULL,
-        post_id INT UNSIGNED NOT NULL,
-        label VARCHAR(100) NOT NULL,
-        KEY thread_id (thread_id)
-      ) ENGINE = InnoDB CHARACTER SET utf8 COLLATE utf8_general_ci
-    ");
-
-    try
+    if ($version == 0)
     {
-      $db->query("ALTER TABLE xf_thread ADD COLUMN has_threadmarks INT UNSIGNED DEFAULT 0 NOT NULL AFTER prefix_id");
+      $db->query("
+        CREATE TABLE IF NOT EXISTS threadmarks (
+          threadmark_id INT UNSIGNED NOT NULL PRIMARY KEY AUTO_INCREMENT,
+          thread_id INT UNSIGNED NOT NULL,
+          post_id INT UNSIGNED NOT NULL,
+          label VARCHAR(100) NOT NULL,
+          UNIQUE KEY `thread_post_id` (`thread_id`,`post_id`)
+        ) ENGINE = InnoDB CHARACTER SET utf8 COLLATE utf8_general_ci
+      ");
+      $tables_created = true;
     }
-    catch (Zend_Db_Exception $e) {}
+
+    if ($version == 1)
+    {
+      if ($db->fetchRow("SHOW COLUMNS FROM xf_thread WHERE Field = ?", 'has_threadmarks'))
+      {
+        $db->query("ALTER TABLE xf_thread DROP COLUMN has_threadmarks");
+      }
+      if ($db->fetchRow("SHOW INDEX FROM threadmarks WHERE Column_name = ?", 'thread_id'))
+      {
+        $db->query("ALTER TABLE threadmarks DROP INDEX thread_id ");
+      }
+    }
+
+    if ($version < 2)
+    {
+      if (!$db->fetchRow("SHOW COLUMNS FROM xf_thread WHERE Field = ?", 'threadmark_count'))
+      {    
+        $db->query("ALTER TABLE xf_thread ADD COLUMN threadmark_count INT UNSIGNED DEFAULT 0 NOT NULL");
+      }
+      if (!$db->fetchRow("SHOW INDEX FROM threadmarks WHERE Key_name = ?", 'thread_post_id'))
+      {
+          $db->query("ALTER TABLE threadmarks add unique index thread_post_id (`thread_id`,`post_id`)");
+      }
+    }
+
+    XenForo_Application::defer('Sidane_Threadmarks_Deferred_Cache', array(), null, true);
   }
 
   public static function uninstall()
   {
     $db = XenForo_Application::get('db');
-    $db->query("ALTER TABLE xf_thread DROP COLUMN has_threadmarks");
-    $db->query("DROP TABLE threadmarks");
+    if ($db->fetchRow("SHOW COLUMNS FROM xf_thread WHERE Field = ?", 'has_threadmarks'))
+    {
+      $db->query("ALTER TABLE xf_thread DROP COLUMN threadmark_count");
+    }  
+    $db->query("DROP TABLE IF EXISTS threadmarks");
   }
 }
-
-?>
